refactor: divided material tests by component type, use should nomenclature.
